cards: Small fixes to the new card decks
Due to an Inkscape bug, the 4-color decks and Rheinland contained
leftovers from the original black-red scheme.
While this is not a problem when using librsvg, QtSVG incorrectly
renders the cards in the original colours. Changing it makes these
card decks compatible as-is with kpat.

Also fixes some positioning issues in the Neoclassical card deck,
as well as increases the size of the suit symbols on the court
cards in Rheinland.

Fixes: https://gitlab.gnome.org/GNOME/aisleriot/-/issues/930